refactored Notes table into smaller components
